backend/pyproject.toml
[project]
name = "newsletter-herald-backend"
version = "0.1.0"
description = "The API Gateway acts as a single entry point that manages client requests and delegates them to the appropriate backend services."
authors = [
    {name = "Anthony A.S Baptiste",email = "anthony.baptiste@outlook.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">3.11"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "requests (>=2.32.4,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "python-docx (>=1.2.0,<2.0.0)",
    "pymupdf (>=1.26.3,<2.0.0)",
    "tiktoken (>=0.9.0,<0.10.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
